requirements.txt
fastapi
uvicorn
numpy
pandas
Pillow
scikit-learn
tensorflow
python-dotenv
jinja2
httpx
python-multipart
motor
tensorflow-hub
spacy
transformers
torch
xgboost
pypdf
scipy
pytesseract
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
fhir.resources
qrcode[pil]
passlib[bcrypt]
python-jose